some more fixes

+++ internal/store/workout_store.go
@@ -35,10 +35,10 @@ func NewPostgresWorkoutStore(db *sql.DB) *PostgresWorkoutStore {
 
 type WorkoutStore interface {
 	CreateWorkout(*Workout) (*Workout, error)
-	GetWorkoutByID(id int64) (*Workout, error)
+	GetWorkoutByID(id int) (*Workout, error)
 	UpdateWorkout(*Workout) error
-	DeleteWorkout(id int64) error
-	GetWorkoutOwner(id int64) (int, error)
+	DeleteWorkout(id int) error
+	GetWorkoutOwner(id int) (int, error)
 	GetWorkoutsForUser(id int) ([]Workout, error)
 }
 
@@ -101,7 +101,7 @@ func (s *PostgresWorkoutStore) CreateWorkout(workout *Workout) (*Workout, error)
 	return workout, nil
 }
 
-func (pg *PostgresWorkoutStore) GetWorkoutByID(id int64) (*Workout, error) {
+func (pg *PostgresWorkoutStore) GetWorkoutByID(id int) (*Workout, error) {
 	workout := &Workout{}
 	query := `
 	SELECT id, title, description, duration_minutes, calories_burned
@@ -209,7 +209,7 @@ func (pg *PostgresWorkoutStore) UpdateWorkout(workout *Workout) error {
 	return tx.Commit()
 }
 
-func (pg *PostgresWorkoutStore) DeleteWorkout(id int64) error {
+func (pg *PostgresWorkoutStore) DeleteWorkout(id int) error {
 	query := `
 	DELETE FROM workouts
 	WHERE id = $1
@@ -231,7 +231,7 @@ func (pg *PostgresWorkoutStore) DeleteWorkout(id int64) error {
 	return nil
 }
 
-func (pg *PostgresWorkoutStore) GetWorkoutOwner(workoutID int64) (int, error) {
+func (pg *PostgresWorkoutStore) GetWorkoutOwner(workoutID int) (int, error) {
 	var userID int
 
 	query := `
@@ -248,7 +248,7 @@ func (pg *PostgresWorkoutStore) GetWorkoutOwner(workoutID int64) (int, error) {
 	return userID, nil
 }
 
-func (pg *PostgresWorkoutStore) GetWorkoutsForUser(userID int64) ([]Workout, error) {
+func (pg *PostgresWorkoutStore) GetWorkoutsForUser(userID int) ([]Workout, error) {
 	query := `
     SELECT id, user_id, title, description, duration_minutes, calories_burned, created_at, updated_at 
     FROM workouts
